Update [*]: global renaming
