+++ src/chunk-factory.ts
@@ -1,4 +1,4 @@
-import { chunkGeneratorFactory } from './generateChunk';
+import { chunkGeneratorFactory } from './generate-chunk';
 import { ChunkImporter, ImportFactory, IChunkPlugin, Chunk } from './types';
 
 export function chunkImporterFactory<C extends Chunk, I extends ImportFactory<C> = ImportFactory<C>>(
@@ -6,7 +6,7 @@ export function chunkImporterFactory<C extends Chunk, I extends ImportFactory<C>
   basePath: string,
   plugins: IChunkPlugin<C>[] = []
 ): ChunkImporter<C> {
-  const generateChunk = chunkGeneratorFactory<C, I>(importFactory, plugins);
+  const generateChunk = chunkGeneratorFactory<I, C>(importFactory, plugins);
 
   return function chunkImporter(name: string, relativePath?: string) {
     const generate = generateChunk(name);
